Update user tests

+++ test/resolver-user.js
@@ -1,7 +1,7 @@
 const appConfig = require('../config');
-const { doesNotMatch } = require('assert');
+//const { doesNotMatch } = require('assert');
 const { expect } = require('chai');
-const fs = require('fs');
+//const fs = require('fs');
 const mongoose = require("mongoose");
 
 const graphqlResolver = require('../graphql/resolvers/user');
@@ -19,10 +19,13 @@ const testUserInfo = {
 	updatedAt: '',
 };
 
+/*global describe, before, after, it */
+
 describe('User functions', () => {
 	before(async () => {
 		await mongoose.connect('mongodb://' + appConfig.db.hostname + ':' + appConfig.db.port + '/' + appConfig.db.name,
 			{ useNewUrlParser: true, useUnifiedTopology: true, useCreateIndex: true });
+		await User.deleteMany();
 	});
 
 	after(async () => {
@@ -39,17 +42,60 @@ describe('User functions', () => {
 			lastName: testUserInfo.lastName,
 		};
 
-		result = await graphqlResolver.RootMutation.createUser(_, { userInput }, req);
+		var result = await graphqlResolver.RootMutation.createUser(_, { userInput }, req);
 		const testUserId = result._id.toString();
-		// cleanup
-		await User.deleteOne({ _id: testUserId });
 
-		expect(result.email).to.equal(userInput.email);
-		expect(result.username).to.equal(userInput.username);
-		expect(result.firstName).to.equal(userInput.firstName);
-		expect(result.lastName).to.equal(userInput.lastName);
-		expect(result.password).to.not.equal(userInput.password);
+		// Don't assume that what is returned is what was stored in the database
+		result = await User.findById(testUserId);
 
+		expect(result).not.to.be.null;
+		expect(result._id.toString()).to.equal(testUserId);
+		if (!testUserId) {
+
+			// cleanup
+			await User.deleteOne({ _id: testUserId });
+
+			expect(result.email).to.equal(userInput.email);
+			expect(result.username).to.equal(userInput.username);
+			expect(result.firstName).to.equal(userInput.firstName);
+			expect(result.lastName).to.equal(userInput.lastName);
+			expect(result.password).to.not.equal(userInput.password);
+		}
+	});
+
+	it('updateUser should update an existing user', async () => {
+		var _, req = {};
+		var userInput = {
+			email: 'updateUser_' + testUserInfo.email,
+			username: 'updateUser_' + testUserInfo.username,
+			password: testUserInfo.password,
+			firstName: testUserInfo.firstName,
+			lastName: testUserInfo.lastName,
+		};
+		const user = new User(userInput);
+		await user.save();
+		const testUserId = user._id.toString();
+
+		userInput.email = 'updateUser2_' + testUserInfo.email;
+		userInput.username = 'updateUser_' + testUserInfo.username + '2';
+		userInput.password = testUserInfo.password + '2';
+		userInput.firstName = testUserInfo.firstName + '2';
+		userInput.lastName = testUserInfo.lastName + '2';
+
+		var result = await graphqlResolver.RootMutation.updateUser(_, { _id: testUserId, userInput }, req);
+
+		expect(result).not.to.be.null;
+		expect(result._id.toString()).to.equal(testUserId);
+		if (!testUserId) {
+			// cleanup
+			await User.deleteOne({ _id: testUserId });
+
+			expect(result.email).to.equal(userInput.email);
+			expect(result.username).to.equal(userInput.username);
+			expect(result.firstName).to.equal(userInput.firstName);
+			expect(result.lastName).to.equal(userInput.lastName);
+			expect(result.password).to.not.equal(userInput.password);
+		}
 	});
 
 	it('getUser should get a user by id', async () => {
@@ -66,7 +112,7 @@ describe('User functions', () => {
 		const testUserId = dbuser._id.toString();
 
 		var _, req = {};
-		result = await graphqlResolver.RootQuery.getUser(_, { _id: testUserId }, req);
+		var result = await graphqlResolver.RootQuery.getUser(_, { _id: testUserId }, req);
 		// cleanup
 		await User.deleteOne({ _id: testUserId });
 
@@ -74,6 +120,26 @@ describe('User functions', () => {
 		expect(result.username).to.equal('getUser_' + testUserInfo.username);
 		expect(result.firstName).to.equal(testUserInfo.firstName);
 		expect(result.lastName).to.equal(testUserInfo.lastName);
+	});
+
+	it('deleteUser should delete a user by id', async () => {
+		// create the test user
+		const user = new User({
+			email: 'deleteUser_' + testUserInfo.email,
+			username: 'deleteUser_' + testUserInfo.username,
+			password: testUserInfo.password,
+			firstName: testUserInfo.firstName,
+			lastName: testUserInfo.lastName,
+		});
+		await user.save();
+		const testUserId = user._id.toString();
+
+		var _, req = {};
+		var result = await graphqlResolver.RootMutation.deleteUser(_, { _id: testUserId }, req);
+		result = await User.findById(testUserId);
+		expect(result).to.be.null;
+		result = await User.findOne({ email: 'deleteUser_' + testUserInfo.email });
+		expect(result).to.be.null;
 	});
 
 });
